remove regolith information, update macos installation script, replace jenv with sdkman

+++ installMac.sh
@@ -1,33 +1,40 @@
 #!/bin/bash
 
-/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+if ! which brew > /dev/null; then
+    /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+fi
+
+# SDKMAN installation
+if ! which sdk > /dev/null; then
+    curl -s "https://get.sdkman.io" | bash
+    source "$HOME/.sdkman/bin/sdkman-init.sh"
+fi
 
 # JAVA
-JAVA_JDK11=javajdk.pkg
-JAVA_JDK17=javajdk17.pkg
+if ! which java > /dev/null; then
+    JAVA_JDK11=javajdk.pkg
+    JAVA_JDK17=javajdk17.pkg
 
-curl -o $JAVA_JDK11 -LO https://corretto.aws/downloads/latest/amazon-corretto-11-x64-macos-jdk.pkg
-curl -o $JAVA_JDK17 -LO https://corretto.aws/downloads/latest/amazon-corretto-17-x64-linux-jdk.tar.gz
-sudo installer -pkg ./$JAVA_JDK11 -target /
-sudo installer -pkg ./$JAVA_JDK17 -target /
-rm $JAVA_JDK11 $JAVA_JDK17
+    curl -o $JAVA_JDK11 -LO https://corretto.aws/downloads/latest/amazon-corretto-11-x64-macos-jdk.pkg
+    curl -o $JAVA_JDK17 -LO https://corretto.aws/downloads/latest/amazon-corretto-17-x64-macos-jdk.pkg
+    
+    sudo installer -pkg ./$JAVA_JDK11 -target /
+    sudo installer -pkg ./$JAVA_JDK17 -target /
+    rm $JAVA_JDK11 $JAVA_JDK17
+
+    sdk install java 11-amzn /Library/Java/JavaVirtualMachines/amazon-corretto-11.jdk/Contents/Home
+    sdk install java 17-amzn /Library/Java/JavaVirtualMachines/amazon-corretto-17.jdk/Contents/Home
+    sdk default java 17-amzn
+fi
 
 # Terminal utils
-brew install mpg123
-brew install tree
-brew install ncdu
-brew install stow
+brew install mpg123 tree ncdu stow
 
 # fzf
 git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
 ~/.fzf/install --all
 
-brew install jenv
-jenv enable-plugin export
-jenv add /Library/Java/JavaVirtualMachines/amazon-corretto-11.jdk/Contents/Home
-jenv add /Library/Java/JavaVirtualMachines/amazon-corretto-17.jdk/Contents/Home
-
-brew install nvm
+#brew install nvm
 
 # INTELLIJ IDEA CE
 #brew install --cask intellij-idea-ce
@@ -45,13 +52,13 @@ brew install --cask fork
 # brew install virtualbox
 
 # Zim
-wget -nv -O - https://raw.githubusercontent.com/zimfw/install/master/install.zsh | zsh
+curl -fsSL https://raw.githubusercontent.com/zimfw/install/master/install.zsh | zsh
 
 # CONFIG
 
-export _git_email=aitor.escolar@galpsolar.com
-export _git_user=aitor.escolar
+export _git_email=aitorescolarcabeza@gmail.com
+export _git_user=aescolar
 export _platform_folder=macos
 
-sh ./scripts/setupGit
-sh ./scripts/setupLinks
+sh ./scripts/setupGit.sh
+sh ./scripts/setupLinks.sh
